backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
beanie==1.26.0
pymongo==4.7.2
redis==5.0.4
celery==5.4.0
httpx==0.27.0
websockets==12.0
web3==6.19.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
pytest-asyncio>=0.23